Extract scheduled email functions to their own subclass
